[FEAT] 정 부분 계산하는 로직 추가
- 가이드와 투어리스트 리뷰에 따른 정 업데이트 로직을 구현함

+++ src/main/java/com/koing/server/koing_server/domain/review/ReviewToGuide.java
@@ -49,7 +49,7 @@ public class ReviewToGuide extends AuditingTimeEntity {
     @Column(nullable = false, name = "guide_reviews")
     private Set<String> guideReviews;
 
-    private int attachment;
+    private double attachment;
 
     @Column(length = 500)
     private String totalReview;
